requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
uvicorn 
python-dotenv